[refactor]DefaultHeaderを別ファイルに移動
