+++ app/src/main/java/com/example/pokebook/ui/screen/DefaultHeaderScreen.kt
@@ -1,0 +1,124 @@
+package com.example.pokebook.ui.screen
+
+import android.util.Log
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.AlertDialog
+import androidx.compose.material.Text
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.pokebook.R
+import com.example.pokebook.ui.viewModel.Search.SearchConditionState
+
+/**
+ * ポケモン一覧のヘッダー領域
+ */
+@Composable
+fun DefaultHeader(
+    title: String,
+    displayPage: Int=0,
+    maxPage: String = "",
+    updateButtonStates: (Boolean, Boolean) -> Unit = { _, _ -> },
+    onClickNext: () -> Unit = {},
+    onClickBack: () -> Unit = {},
+) {
+    Column(
+        modifier = Modifier
+            .background(MaterialTheme.colorScheme.background)
+            .padding(5.dp)
+            .fillMaxWidth(),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Text(
+            text = title,
+            color = MaterialTheme.colorScheme.onBackground,
+            fontSize = 25.sp,
+            modifier = Modifier
+                .padding(2.dp)
+                .align(Alignment.CenterHorizontally)
+                .background(MaterialTheme.colorScheme.background),
+            textAlign = TextAlign.Center
+        )
+        Row(
+            modifier = Modifier
+                .padding(vertical = 6.dp)
+        ) {
+            Log.d("test","displayPage:$displayPage")
+            if (displayPage == 0) {
+                Spacer(modifier = Modifier.weight(1F))
+            } else {
+                Box(
+                    modifier = Modifier
+                        .weight(1F)
+                        .wrapContentHeight()
+                        .clickable {
+                            updateButtonStates.invoke(true, false)
+                            onClickBack.invoke()
+                        },
+                    contentAlignment = Alignment.CenterEnd
+                ) {
+                    Text(
+                        text = stringResource(R.string.back_button),
+                        color = MaterialTheme.colorScheme.onBackground,
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .border(
+                                width = 1.dp,
+                                color = Color.DarkGray,
+                                shape = RoundedCornerShape(20.dp)
+                            )
+                            .padding(4.dp),
+                        textAlign = TextAlign.Center
+                    )
+                }
+            }
+            Box(
+                modifier = Modifier
+                    .weight(1F)
+                    .wrapContentHeight()
+                    .clickable {
+                        updateButtonStates.invoke(false, true)
+                        onClickNext.invoke()
+                    },
+                contentAlignment = Alignment.CenterEnd
+            ) {
+                Text(
+                    text = stringResource(R.string.next_button),
+                    color = MaterialTheme.colorScheme.onBackground,
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .border(
+                            width = 1.dp,
+                            color = Color.DarkGray,
+                            shape = RoundedCornerShape(20.dp)
+                        )
+                        .padding(4.dp),
+                    textAlign = TextAlign.Center
+                )
+            }
+        }
+    }
+}
