+++ app/src/main/java/com/example/pokebook/ui/screen/HomeScreen.kt
@@ -260,89 +260,6 @@ fun PokeCard(
     }
 }
 
-/**
- * ポケモン一覧のヘッダー領域
- */
-@Composable
-fun DefaultHeader(
-    title: String,
-    updateButtonStates: (Boolean, Boolean) -> Unit = { _, _ -> },
-    onClickNext: () -> Unit = {},
-    onClickBack: () -> Unit = {},
-) {
-    Column(
-        modifier = Modifier
-            .background(MaterialTheme.colorScheme.background)
-            .padding(5.dp)
-            .fillMaxWidth(),
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(
-            text = title,
-            color = MaterialTheme.colorScheme.onBackground,
-            fontSize = 25.sp,
-            modifier = Modifier
-                .padding(2.dp)
-                .align(Alignment.CenterHorizontally)
-                .background(MaterialTheme.colorScheme.background),
-            textAlign = TextAlign.Center
-        )
-        Row(
-            modifier = Modifier
-                .padding(vertical = 6.dp)
-        ) {
-            Box(
-                modifier = Modifier
-                    .weight(1F)
-                    .wrapContentHeight()
-                    .clickable {
-                        updateButtonStates.invoke(true, false)
-                        onClickBack.invoke()
-                    },
-                contentAlignment = Alignment.CenterEnd
-            ) {
-                Text(
-                    text = stringResource(R.string.back_button),
-                    color = MaterialTheme.colorScheme.onBackground,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .border(
-                            width = 1.dp,
-                            color = Color.DarkGray,
-                            shape = RoundedCornerShape(20.dp)
-                        )
-                        .padding(4.dp),
-                    textAlign = TextAlign.Center
-                )
-            }
-            Box(
-                modifier = Modifier
-                    .weight(1F)
-                    .wrapContentHeight()
-                    .clickable {
-                        updateButtonStates.invoke(false, true)
-                        onClickNext.invoke()
-                    },
-                contentAlignment = Alignment.CenterEnd
-            ) {
-                Text(
-                    text = stringResource(R.string.next_button),
-                    color = MaterialTheme.colorScheme.onBackground,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .border(
-                            width = 1.dp,
-                            color = Color.DarkGray,
-                            shape = RoundedCornerShape(20.dp)
-                        )
-                        .padding(4.dp),
-                    textAlign = TextAlign.Center
-                )
-            }
-        }
-    }
-}
-
 @Preview
 @Composable
 private fun PokeCardPreview() {
